vm: split ops.rs into opcode-family modules (#52)
The 1628-line `crates/harn-vm/src/vm/ops.rs` held every bytecode
handler in a single `impl Vm` block. Split it into `ops/mod.rs` plus
12 per-family files, each under 413 lines. Dispatch uses a chain of
`try_execute_<family>_op(op) -> Result<bool, VmError>` helpers in
`ops/mod.rs`: each returns `true` when it handled the opcode, `false`
to fall through to the next family. Behavior is unchanged — same
stack ops, same side effects, same error propagation (`VmError::Return`
still unwinds from the `Return` handler in `call.rs`).

Layout:
- stack.rs: Constant/Nil/True/False, GetVar/DefLet/DefVar/SetVar,
  PushScope/PopScope, Pop/Dup/Swap, GetArgc
- arithmetic.rs: Add/Sub/Mul/Div/Mod/Pow/Negate (+ private helpers)
- comparison.rs: Equal/NotEqual/Less/Greater/LessEqual/GreaterEqual/Contains
- logical.rs: Not
- control_flow.rs: Jump/JumpIfFalse/JumpIfTrue
- call.rs: Call/CallSpread/TailCall/Return/Closure/MethodCall{,Opt,Spread}/Pipe
- collections.rs: BuildList/BuildDict, Subscript/Slice, Get/Set Property,
  SetSubscript, Concat, BuildEnum/MatchEnum
- iter.rs: IterInit/IterNext/PopIterator
- parallel.rs: Parallel/ParallelMap/ParallelSettle/Spawn/Deadline* +
  parallel_cap_from_value / run_capped_ordered helpers
- exception.rs: Throw/TryCatchSetup/PopHandler/TryUnwrap
- imports.rs: Import/SelectiveImport
- misc.rs: CheckType/Yield

Verified: cargo check/clippy, 735 harn-vm tests, 505 conformance
tests, full `make all` gate all pass.
